more gear picker cleanup for consistency
