+++ ui/core/components/individual_sim_ui/item_swap_picker.tsx
@@ -2,12 +2,12 @@ import tippy from 'tippy.js';
 
 import { Player } from '../../player.js';
 import { ItemSlot, Spec } from '../../proto/common.js';
-import { SimUI } from '../../sim_ui.js';
+import { SimUI } from '../../sim_ui.jsx';
 import { EventID, TypedEvent } from '../../typed_event.js';
 import { BooleanPicker } from '../boolean_picker.js';
 import { Component } from '../component.js';
+import { IconItemSwapPicker } from '../gear_picker/icon_item_swap_picker.jsx';
 import { Input } from '../input.jsx';
-import { IconItemSwapPicker } from './icon_item_swap_picker.jsx';
 
 export interface ItemSwapConfig {
 	itemSlots: Array<ItemSlot>;
