+++ ui/core/components/individual_sim_ui/settings_tab.ts
@@ -18,13 +18,13 @@ import * as IconInputs from '../icon_inputs';
 import { Input } from '../input';
 import * as BuffDebuffInputs from '../inputs/buffs_debuffs';
 import { relevantStatOptions } from '../inputs/stat_options';
-import { ItemSwapPicker } from '../item_swap/item_swap_picker';
 import { MultiIconPicker, MultiIconPickerItemConfig } from '../multi_icon_picker';
 import { NumberPicker } from '../number_picker';
 import { SavedDataManager } from '../saved_data_manager';
 import { SimTab } from '../sim_tab';
 import { IsbConfig } from './../other_inputs';
 import { ConsumesPicker } from './consumes_picker';
+import { ItemSwapPicker } from './item_swap_picker';
 import { PresetBuildsPicker } from './preset_builds_picker';
 
 export class SettingsTab extends SimTab {
