+++ ui/core/individual_sim_ui.ts
@@ -7,11 +7,11 @@ import * as IconInputs from './components/icon_inputs';
 import * as Importers from './components/importers';
 import { BulkTab } from './components/individual_sim_ui/bulk_tab';
 import { GearTab } from './components/individual_sim_ui/gear_tab';
+import { ItemSwapConfig } from './components/individual_sim_ui/item_swap_picker';
 import { RotationTab } from './components/individual_sim_ui/rotation_tab';
 import { SettingsTab } from './components/individual_sim_ui/settings_tab';
 import { TalentsTab } from './components/individual_sim_ui/talents_tab';
 import * as InputHelpers from './components/input_helpers';
-import { ItemSwapConfig } from './components/item_swap/item_swap_picker';
 import { addRaidSimAction, RaidSimResultsManager } from './components/raid_sim_action';
 import { SavedDataConfig } from './components/saved_data_manager';
 import { addStatWeightsAction } from './components/stat_weights_action';
